feat(rs)!: Split `Value` into hyperedge values and function i/o (#2)
`Value`s are used for two different cases in the format,
- To identify hyperedges in a ValueTable, with a valid `ValueId`.
- To list function input/outputs, without an associated Id.

This is fine at the format level, but annoying when using an API and
having to `unwrap` accesses to the ids that we know should be there.

This PR splits the wrapper type into two, so we have type-safe access to
hyperedge ids.

BREAKING CHANGE: Split the `Value` definition into `FunctionIOValue` and
`WireValue`.