Add custom Python bindings for jeff (#3)
This PR adds custom Python bindings for jeff around the `pycapnp`
package, which significantly improves ease-of-use for both reading and
writing jeff from Python, as well as being able to provide additional
verification & documentation.

There are some things left to do, like adding the rest of the op
instructions (mostly mechanical work), adding support for metadata, and
potential additional simplifications (see comments in the module). The
class structure for all core schema structs is in-place however, with
dual-purpose reading/writing objects as well as a partially flattened
struct hierarchy.