feat(py): Setup python bindings project (#8)
Setups a python project under `impl/py`, so we can publish the
`jeff-format` package.

There are some problems with setting up dependencies, but we can merge
this as-is for the moment.

The code from #3 is left unchanged apart from adding the capnp schema to
a sub-dir and running automatic linters.