chore: RFC process guideline and template (#31)
Defines the process for proposing RFCs for the Jeff project.

This is based on the Rust RFC guidelines. See
[rust-lang/rfcs](https://github.com/rust-lang/rfcs).
I simplified the RFC lifetime steps, to avoid burdening the process too
much at the begining.

While the guideline can be updated over time, we should try to define
now the kind of proposals we require follow this process. Some
categories we may consider:

- Changes to the Jeff specification that are broader than a bugfix or
documentation improvement.
- Reorganizations of the Jeff sub-projects / libraries.
- Bugfixes and documentation changes to the specification (that trigger
a new document version)
- Changes to the project governance.
- Changes to specific language libraries.

At time of creation, this PR only proposes including the first two
categories. (See
[`#when-you-need-to-follow-this-process`](https://github.com/unitaryfoundation/jeff/blob/42fe49a32e2efe5c50f1bac823ef41fbf6df26df/rfcs/README.md#when-you-need-to-follow-this-process)).

The template is mostly geared towards schema/implementation changes, but
it should be adaptable for the other cases.

---------

Co-authored-by: Josh Izaac <josh146@gmail.com>